refactor: streamline llm models
Based on a lengthy offline discussion:
- There are no longer OpenAI/Google/Anthropic models, just the base LLM
  class
- The LLM class methods don't take params to avoid potential precedence
  confusion between params passed in at call time vs params set on the
  model. If you are using a model override, it overrides all params, and
  user can take responsibility for managing that with user-level code
  that shares default parameters across calls and overrides.
- We swap the order of client and params for the model

I renamed models/base.py to models/llm.py because it's no longer a base,
and I prefer when single-class files have filenames matching
the class.
